PHOENIX-5721 Unify 4.x branches (HBase 1.4 and 1.5)
adds the build profile infrastructure
factors out the HBase 1.4 and 1.5 compatibility layers
