--- build.txt
@@ -1,64 +1,0 @@
-############################################################################
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-############################################################################
-
-
-# Building Apache Phoenix
-=========================
-
-Phoenix uses Maven (3.X) to build all its necessary resources. 
-
-## Building from source
-=======================
-
-On first setup, you may need to run
-	$ mvn install -DskipTests
-to install the local jars. This is a side-effect of multi-module maven projects
-
-1. To re-generate the antlr based files:
-	$ mvn process-sources
-
-2. To build the jars
-	$ mvn package
-and optionally, to just skip all the tests and build the jars:
-	$ mvn package -DskipTests
-
-Note: javadocs are generated in target/apidocs
-
-## Importing into eclipse
-=========================
-
-Use the m2e eclipse plugin and do Import->Maven Project and just pick the root 'phoenix' directory.
-
-## Running the tests
-====================
-
-1. All tests 
-	$ mvn clean test
-
-## Findbugs
-===========
-Findbugs report is generated in /target/site
-	$ mvn site
-	
-
-## Generate Apache Web Site
-===========================
-checkout https://svn.apache.org/repos/asf/phoenix
-        $ build.sh
