+++ phoenix-core/src/main/java/org/apache/phoenix/coprocessor/DelegateRegionObserver.java
@@ -61,15 +61,14 @@ import org.apache.hadoop.hbase.regionserver.wal.WALEdit;
 import org.apache.hadoop.hbase.security.User;
 import org.apache.hadoop.hbase.util.Pair;
 import org.apache.hadoop.hbase.wal.WALKey;
+import org.apache.phoenix.compat.hbase.CompatDelegateRegionObserver;
 
 import com.google.common.collect.ImmutableList;
 
-public class DelegateRegionObserver implements RegionObserver {
-
-    protected final RegionObserver delegate;
+public class DelegateRegionObserver extends CompatDelegateRegionObserver implements RegionObserver {
     
     public DelegateRegionObserver(RegionObserver delegate) {
-        this.delegate = delegate;
+        super(delegate);
     }
 
     @Override
@@ -679,37 +678,10 @@ public class DelegateRegionObserver implements RegionObserver {
         return delegate.postInstantiateDeleteTracker(ctx, delTracker);
     }
 
-    @Override
+    //No @Override for backwards compatibility reasons
     public void preWALAppend(ObserverContext<RegionCoprocessorEnvironment> ctx, WALKey key,
             WALEdit edit) throws IOException {
     }
 
-    @Override
-    public InternalScanner preFlushScannerOpen(ObserverContext<RegionCoprocessorEnvironment> c,
-            Store store, KeyValueScanner memstoreScanner, InternalScanner s, long readPoint)
-            throws IOException {
-        return delegate.preFlushScannerOpen(c, store, memstoreScanner, s, readPoint);
-    }
-
-    @Override
-    public InternalScanner preCompactScannerOpen(ObserverContext<RegionCoprocessorEnvironment> c,
-            Store store, List<? extends KeyValueScanner> scanners, ScanType scanType,
-            long earliestPutTs, InternalScanner s, CompactionRequest request, long readPoint)
-            throws IOException {
-      return delegate.preCompactScannerOpen(c, store, scanners, scanType, earliestPutTs, s, request,
-          readPoint);
-    }
-
-    @Override
-    public void preCommitStoreFile(ObserverContext<RegionCoprocessorEnvironment> ctx, byte[] family,
-            List<Pair<Path, Path>> pairs) throws IOException {
-        delegate.preCommitStoreFile(ctx, family, pairs);
-    }
-
-    @Override
-    public void postCommitStoreFile(ObserverContext<RegionCoprocessorEnvironment> ctx,
-            byte[] family, Path srcPath, Path dstPath) throws IOException {
-        delegate.postCommitStoreFile(ctx, family, srcPath, dstPath);
-    }
 
 }
