+++ phoenix-core/src/main/java/org/apache/phoenix/coprocessor/PhoenixAccessController.java
@@ -58,6 +58,7 @@ import org.apache.hadoop.hbase.security.access.*;
 import org.apache.hadoop.hbase.security.access.Permission.Action;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.zookeeper.ZooKeeperWatcher;
+import org.apache.phoenix.compat.hbase.CompatObserverContext;
 import org.apache.phoenix.coprocessor.PhoenixMetaDataCoprocessorHost.PhoenixMetaDataControllerEnvironment;
 import org.apache.phoenix.query.QueryServices;
 import org.apache.phoenix.query.QueryServicesOptions;
@@ -171,7 +172,9 @@ public class PhoenixAccessController extends BaseMetaDataEndpointObserver {
                 htd.addFamily(new HColumnDescriptor(familyName));
             }
             for (BaseMasterAndRegionObserver observer : getAccessControllers()) {
-                observer.preCreateTable(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()), htd, null);
+                observer.preCreateTable(
+                    new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()), htd,
+                    null);
             }
         }
 
@@ -358,12 +361,15 @@ public class PhoenixAccessController extends BaseMetaDataEndpointObserver {
 
         for (BaseMasterAndRegionObserver observer : getAccessControllers()) {
             if (tableType != PTableType.VIEW) {
-                observer.preDeleteTable(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()), physicalTableName);
+                observer.preDeleteTable(
+                    new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
+                    physicalTableName);
             }
             if (indexes != null) {
                 for (PTable index : indexes) {
-                    observer.preDeleteTable(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
-                            TableName.valueOf(index.getPhysicalName().getBytes()));
+                    observer.preDeleteTable(
+                        new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
+                        TableName.valueOf(index.getPhysicalName().getBytes()));
                 }
             }
         }
@@ -383,8 +389,9 @@ public class PhoenixAccessController extends BaseMetaDataEndpointObserver {
         if (!accessCheckEnabled) { return; }
         for (BaseMasterAndRegionObserver observer : getAccessControllers()) {
             if (tableType != PTableType.VIEW) {
-                observer.preModifyTable(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()), physicalTableName,
-                    new HTableDescriptor(physicalTableName));
+                observer.preModifyTable(
+                    new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
+                    physicalTableName, new HTableDescriptor(physicalTableName));
             }
         }
         if (tableType == PTableType.VIEW) {
@@ -401,8 +408,9 @@ public class PhoenixAccessController extends BaseMetaDataEndpointObserver {
             throws IOException {
         if (!accessCheckEnabled) { return; }
         for (BaseMasterAndRegionObserver observer : getAccessControllers()) {
-            observer.preListNamespaceDescriptors(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
-                    Arrays.asList(NamespaceDescriptor.create(schemaName).build()));
+            observer.preListNamespaceDescriptors(
+                new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
+                Arrays.asList(NamespaceDescriptor.create(schemaName).build()));
         }
     }
 
@@ -411,8 +419,9 @@ public class PhoenixAccessController extends BaseMetaDataEndpointObserver {
             throws IOException {
         if (!accessCheckEnabled) { return; }
         for (BaseMasterAndRegionObserver observer : getAccessControllers()) {
-            observer.preCreateNamespace(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
-                    NamespaceDescriptor.create(schemaName).build());
+            observer.preCreateNamespace(
+                new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
+                NamespaceDescriptor.create(schemaName).build());
         }
     }
 
@@ -421,7 +430,9 @@ public class PhoenixAccessController extends BaseMetaDataEndpointObserver {
             throws IOException {
         if (!accessCheckEnabled) { return; }
         for (BaseMasterAndRegionObserver observer : getAccessControllers()) {
-            observer.preDeleteNamespace(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()), schemaName);
+            observer.preDeleteNamespace(
+                new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
+                schemaName);
         }
     }
 
@@ -431,8 +442,9 @@ public class PhoenixAccessController extends BaseMetaDataEndpointObserver {
             throws IOException {
         if (!accessCheckEnabled) { return; }
         for (BaseMasterAndRegionObserver observer : getAccessControllers()) {
-            observer.preModifyTable(new ObserverContext<MasterCoprocessorEnvironment>(getActiveUser()), physicalTableName,
-                    new HTableDescriptor(physicalTableName));
+            observer.preModifyTable(
+                new CompatObserverContext<MasterCoprocessorEnvironment>(getActiveUser()),
+                physicalTableName, new HTableDescriptor(physicalTableName));
         }
         // Check for read access in case of rebuild
         if (newState == PIndexState.BUILDING) {
