+++ phoenix-core/src/main/java/org/apache/phoenix/coprocessor/PhoenixMetaDataCoprocessorHost.java
@@ -32,12 +32,13 @@ import org.apache.hadoop.hbase.coprocessor.CoprocessorHost;
 import org.apache.hadoop.hbase.coprocessor.ObserverContext;
 import org.apache.hadoop.hbase.coprocessor.RegionCoprocessorEnvironment;
 import org.apache.hadoop.hbase.ipc.RpcServer;
-import org.apache.hadoop.hbase.metrics.MetricRegistry;
 import org.apache.hadoop.hbase.regionserver.Region;
 import org.apache.hadoop.hbase.regionserver.RegionCoprocessorHost;
 import org.apache.hadoop.hbase.regionserver.RegionServerServices;
 import org.apache.hadoop.hbase.security.User;
 import org.apache.hadoop.hbase.security.UserProvider;
+import org.apache.phoenix.compat.hbase.CompatObserverContext;
+import org.apache.phoenix.compat.hbase.CompatPhoenixMetaDataControllerEnvironment;
 import org.apache.phoenix.query.QueryServices;
 import org.apache.phoenix.query.QueryServicesOptions;
 import org.apache.phoenix.schema.PIndexState;
@@ -66,7 +67,8 @@ public class PhoenixMetaDataCoprocessorHost
         loadSystemCoprocessors(conf, PHOENIX_META_DATA_COPROCESSOR_CONF_KEY);
     }
 
-    private static abstract class CoprocessorOperation<T extends CoprocessorEnvironment> extends ObserverContext<T> {
+    private static abstract class CoprocessorOperation<T extends CoprocessorEnvironment>
+            extends CompatObserverContext<T> {
         abstract void call(MetaDataEndpointObserver oserver, ObserverContext<T> ctx) throws IOException;
 
         public CoprocessorOperation(User user) {
@@ -115,15 +117,13 @@ public class PhoenixMetaDataCoprocessorHost
     /**
      * Encapsulation of the environment of each coprocessor
      */
-    public static class PhoenixMetaDataControllerEnvironment extends CoprocessorHost.Environment
+    public static class PhoenixMetaDataControllerEnvironment 
+            extends CompatPhoenixMetaDataControllerEnvironment
             implements RegionCoprocessorEnvironment {
-
-        private RegionCoprocessorEnvironment env;
 
         PhoenixMetaDataControllerEnvironment(RegionCoprocessorEnvironment env, Coprocessor instance,
                 int priority, int sequence, Configuration conf) {
-            super(instance, priority, sequence, conf);
-            this.env = env;
+            super(env, instance, priority, sequence, conf);
         }
 
         @Override
@@ -150,10 +150,6 @@ public class PhoenixMetaDataCoprocessorHost
             return env.getSharedData();
         }
 
-        @Override
-        public MetricRegistry getMetricRegistryForRegionServer() {
-            return env.getMetricRegistryForRegionServer();
-        }
     }
 
     @Override
