+++ phoenix-hbase-compat-1.5.0/src/main/java/org/apache/phoenix/compat/hbase/CompatPhoenixMetaDataControllerEnvironment.java
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.phoenix.compat.hbase;
+
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.hbase.Coprocessor;
+import org.apache.hadoop.hbase.coprocessor.CoprocessorHost;
+import org.apache.hadoop.hbase.coprocessor.RegionCoprocessorEnvironment;
+import org.apache.hadoop.hbase.metrics.MetricRegistry;
+
+public abstract class CompatPhoenixMetaDataControllerEnvironment extends CoprocessorHost.Environment
+        implements RegionCoprocessorEnvironment {
+
+    protected RegionCoprocessorEnvironment env;
+
+    public CompatPhoenixMetaDataControllerEnvironment(RegionCoprocessorEnvironment env,
+            Coprocessor instance, int priority, int sequence, Configuration conf) {
+        super(instance, priority, sequence, conf);
+        this.env = env;
+    }
+
+    @Override
+    public MetricRegistry getMetricRegistryForRegionServer() {
+        return env.getMetricRegistryForRegionServer();
+    }
+}
