+++ phoenix-pherf/pom.xml
@@ -21,7 +21,7 @@
 	<parent>
 		<groupId>org.apache.phoenix</groupId>
 		<artifactId>phoenix</artifactId>
-		<version>4.16.0-HBase-1.5-SNAPSHOT</version>
+		<version>4.16.0-SNAPSHOT</version>
 	</parent>
 
 	<artifactId>phoenix-pherf</artifactId>
