+++ pom.xml
@@ -20,7 +20,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>org.apache.phoenix</groupId>
   <artifactId>phoenix</artifactId>
-  <version>4.16.0-HBase-1.5-SNAPSHOT</version>
+  <version>4.16.0-SNAPSHOT</version>
   <packaging>pom</packaging>
   <name>Apache Phoenix</name>
   <description>A SQL layer over HBase</description>
@@ -40,6 +40,8 @@
   </organization>
 
   <modules>
+    <module>phoenix-hbase-compat-1.5.0</module>
+    <module>phoenix-hbase-compat-1.4.0</module>
     <module>phoenix-core</module>
     <module>phoenix-pherf</module>
     <module>phoenix-client</module>
@@ -72,15 +74,19 @@
   </scm>
 
   <properties>
+    <!-- The HBase compatibility module that that will be included in
+    the shaded JARs and the assembly -->
+    <hbase.profile>1.5</hbase.profile>
+    <hbase.profile.string>hbase-${hbase.profile}</hbase.profile.string>
+
+    <!-- Hadoop Versions -->
+    <hadoop-two.version>2.7.5</hadoop-two.version>
+
     <!-- General Properties -->
     <antlr-input.dir>src/main/antlr3</antlr-input.dir>
     <antlr-output.dir>target/generated-sources/antlr3</antlr-output.dir>
     <test.output.tofile>true</test.output.tofile>
     <top.dir>${project.basedir}</top.dir>
-
-    <!-- Hadoop Versions -->
-    <hbase.version>1.5.0</hbase.version>
-    <hadoop-two.version>2.7.5</hadoop-two.version>
 
     <!-- Dependency versions -->
     <commons-cli.version>1.2</commons-cli.version>
@@ -128,6 +134,7 @@
     <!-- Plugin versions -->
     <maven-eclipse-plugin.version>2.9</maven-eclipse-plugin.version>
     <maven-build-helper-plugin.version>1.9.1</maven-build-helper-plugin.version>
+    <maven-enforcer-plugin.version>3.0.0-M3</maven-enforcer-plugin.version>
 
     <!-- Plugin options -->
     <numForkedUT>8</numForkedUT>
@@ -145,6 +152,11 @@
   <build>
     <pluginManagement>
       <plugins>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-enforcer-plugin</artifactId>
+          <version>${maven-enforcer-plugin.version}</version>
+        </plugin>
         <plugin>
           <groupId>org.apache.maven.plugins</groupId>
           <artifactId>maven-compiler-plugin</artifactId>
@@ -371,10 +383,56 @@
           <artifactId>maven-bundle-plugin</artifactId>
           <version>2.5.3</version>
         </plugin>
+        <plugin>
+          <groupId>org.codehaus.mojo</groupId>
+          <artifactId>exec-maven-plugin</artifactId>
+          <version>1.6.0</version>
+        </plugin>
       </plugins>
     </pluginManagement>
 
     <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-enforcer-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>check-hbase-compatibility</id>
+            <phase>validate</phase>
+            <goals>
+              <goal>enforce</goal>
+            </goals>
+            <configuration>
+              <rules>
+                <evaluateBeanshell>
+                 <condition>
+                 <!-- Add check for any know incompatibilies in the below code -->
+                   import java.util.regex.Pattern;
+                   import java.lang.Integer;
+
+                   versionPattern = Pattern.compile("(\\d+)\\.(\\d+)\\.(\\d+)[^.]*$");
+                   versionMatcher = versionPattern.matcher("${hbase.version}");
+                   versionMatcher.find();
+
+                   hbaseMajor = Integer.parseInt(versionMatcher.group(1));
+                   hbaseMinor = Integer.parseInt(versionMatcher.group(2));
+                   hbasePatch = Integer.parseInt(versionMatcher.group(3));
+
+                   hbaseMajor == 1 &amp;&amp; (
+                     ("${hbase.compat.version}".equals("1.5.0")
+                       &amp;&amp; hbaseMinor == 5)
+                     || ("${hbase.compat.version}".equals("1.4.0")
+                       &amp;&amp; hbaseMinor == 4)
+                     || ("${hbase.compat.version}".equals("1.3.0")
+                       &amp;&amp; hbaseMinor == 3)
+                   )
+                   </condition>
+                  </evaluateBeanshell>
+                </rules>
+              </configuration>
+            </execution>
+          </executions>
+        </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-checkstyle-plugin</artifactId>
@@ -523,6 +581,17 @@
         <classifier>tests</classifier>
         <scope>test</scope>
       </dependency>
+      <dependency>
+        <groupId>org.apache.phoenix</groupId>
+        <artifactId>phoenix-hbase-compat-1.5.0</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.phoenix</groupId>
+        <artifactId>phoenix-hbase-compat-1.4.0</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <!-- TODO Remove PQS and connector dependency rules -->
       <dependency>
         <groupId>org.apache.phoenix</groupId>
         <artifactId>phoenix-flume</artifactId>
@@ -1025,7 +1094,7 @@
   </dependencyManagement>
 
   <profiles>
-    <!-- disable doclint with 1.8+ JDKs-->
+    <!-- disable doclint with 1.8+ JDKs -->
     <profile>
       <id>java8-doclint-disabled</id>
       <activation>
@@ -1074,6 +1143,66 @@
         <spark.version>1.6.1</spark.version>
         <scala.version>2.10.4</scala.version>
         <scala.binary.version>2.10</scala.binary.version>
+      </properties>
+    </profile>
+    <!-- See BUILDING.md for profile selection -->
+    <profile>
+      <id>phoenix-hbase-compat-1.5.0-default</id>
+      <activation>
+        <property>
+          <name>!hbase.profile</name>
+        </property>
+      </activation>
+      <dependencies>
+        <dependency>
+          <groupId>org.apache.phoenix</groupId>
+          <artifactId>phoenix-hbase-compat-1.5.0</artifactId>
+        </dependency>
+      </dependencies>
+      <properties>
+        <hbase.profile>1.5</hbase.profile>
+        <hbase.compat.version>1.5.0</hbase.compat.version>
+        <hbase.version>1.5.0</hbase.version>
+      </properties>
+    </profile>
+    <profile>
+      <id>phoenix-hbase-compat-1.5.0</id>
+      <activation>
+        <property>
+          <name>hbase.profile</name>
+          <value>1.5</value>
+        </property>
+      </activation>
+      <dependencies>
+        <dependency>
+          <groupId>org.apache.phoenix</groupId>
+          <artifactId>phoenix-hbase-compat-1.5.0</artifactId>
+        </dependency>
+      </dependencies>
+      <properties>
+        <hbase.profile>1.5</hbase.profile>
+        <hbase.compat.version>1.5.0</hbase.compat.version>
+        <hbase.version>1.5.0</hbase.version>
+      </properties>
+    </profile>
+    <profile>
+      <id>phoenix-hbase-compat-1.4.0</id>
+      <activation>
+        <property>
+          <name>hbase.profile</name>
+          <value>1.4</value>
+        </property>
+      </activation>
+      <dependencies>
+        <dependency>
+          <groupId>org.apache.phoenix</groupId>
+          <artifactId>phoenix-hbase-compat-1.4.0</artifactId>
+        </dependency>
+      </dependencies>
+      <properties>
+        <hbase.profile>1.4</hbase.profile>
+        <hbase.compat.version>1.4.0</hbase.compat.version>
+        <hbase.version>1.4.10</hbase.version>
       </properties>
     </profile>
   </profiles>
